Check Content-Type before unmarshalling json
This addresses the issue #84. WHen the API servers were returned a 502, linodego paniced. This happened due to the client trying to parse the response as json when a `text/html` response was received.

The solution is to check the response content-type to see if it matches the content-type accepted by the request. All of the code base assumes that the response is well-formed json, so it may be okay to just check to see if the content-type matches "application/json".

If you remove the code added in `errors.go` the included test will fail due to panicing when trying to unmarshal the html response.